pyproject.toml
[project]
name = "telegramjob"
version = "0.1.0"
description = "A Telegram bot that automates screenshots."
readme = "README.md"
requires-python = ">=3.11.0"
dependencies = [
    "asyncio>=4.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.0.0",
    "playwright>=1.54.0",
    "python-dotenv>=1.1.1",
    "telethon>=1.40.0",
    "coloredlogs>=15.0.1",
    "pylint>=3.3.8",
]
